refactor: Structure
