+++ src/modules/readme/models/include-options-field.ts
@@ -1,0 +1,7 @@
+import { TaskInputDefinition } from './task-input-definition';
+
+export interface IncludeOptionsField {
+  field: keyof TaskInputDefinition;
+  title: string;
+  align?: 'left' | 'center' | 'right';
+}
