+++ src/modules/readme/models/index.ts
@@ -1,3 +1,8 @@
 export * from './task-definition';
 export * from './task-input-definition';
 export * from './task-input-type';
+export * from './readme-config';
+export * from './required-options';
+export * from './include-options-field';
+export * from './readme-partial';
+export * from './readme-partial-root';
