+++ src/modules/readme/models/readme-partial-root.ts
@@ -1,0 +1,5 @@
+import { ReadmePartial } from './readme-partial';
+
+export interface ReadmePartialRoot {
+  [key: string]: ReadmePartial;
+}
