+++ src/modules/readme/models/readme-partial.ts
@@ -1,0 +1,4 @@
+export interface ReadmePartial {
+  file: string;
+  format?: string;
+}
