+++ src/modules/readme/models/replace-match.ts
@@ -1,0 +1,4 @@
+export interface ReplaceMatch {
+  content?: string;
+  full?: string;
+}
