+++ src/modules/readme/models/required-options.ts
@@ -1,0 +1,4 @@
+export interface RequiredOptions {
+  true: string;
+  false: string;
+}
