+++ src/modules/readme/readme-constats.ts
@@ -1,4 +1,7 @@
+import TaskFieldFormatter from './formatters/task-field-formatter';
+import TaskInputFormatter from './formatters/task-input-formatter';
 import { ReadmeConfig } from './models/readme-config';
+import { ReplacementCommand } from './models/replacement-command';
 
 export const README_NAME = 'readme.json';
 export const README_DEFAULT_FILE: ReadmeConfig = {
@@ -13,3 +16,21 @@ export const README_DEFAULT_FILE: ReadmeConfig = {
     true: ':white_check_mark:'
   }
 };
+
+export const replacementCommands: ReplacementCommand[] = [
+  {
+    command: 'task-input',
+    formatter: () => new TaskInputFormatter(),
+    options: [{ name: 'task' }, { name: 'type' }]
+  },
+  {
+    command: 'task-field',
+    formatter: () => new TaskFieldFormatter(),
+    options: [
+      { name: 'task' },
+      { name: 'field' },
+      { name: 'objectHandle', optional: true },
+      { name: 'codeFormat', optional: true }
+    ]
+  }
+];
