+++ src/modules/readme/readme-service.ts
@@ -1,0 +1,29 @@
+import fs from 'fs/promises';
+import CommandService from './command-service';
+
+export default class ReadmeService {
+  private _cmdService: CommandService;
+  constructor() {
+    this._cmdService = new CommandService();
+  }
+  async processReadMe(filePath: string) {
+    let content = (await fs.readFile(filePath)).toString();
+
+    const matches = this._cmdService.getMatches(content);
+
+    for (const match of matches) {
+      if (match.content?.startsWith('#')) {
+        const commandResult = this._cmdService.getCommand(match.content);
+
+        if (commandResult && match.full) {
+          const formatter = commandResult.command.formatter();
+          const dd = formatter.getOptions(commandResult.options);
+          const res = await formatter.getFormatted(dd);
+          console.log(res);
+          content = content.replace(match.full, res);
+        }
+      }
+    }
+    return content;
+  }
+}
